+++ src/EditorFeatures/Core/Implementation/IntelliSense/IDocumentProvider.cs
@@ -3,9 +3,6 @@
 // See the LICENSE file in the project root for more information.
 
 using System.Threading;
-using Microsoft.CodeAnalysis.Editor.Shared.Extensions;
-using Microsoft.CodeAnalysis.Editor.Shared.Utilities;
-using Microsoft.CodeAnalysis.Text;
 using Microsoft.VisualStudio.Text;
 
 namespace Microsoft.CodeAnalysis.Editor.Implementation.IntelliSense
@@ -13,21 +10,5 @@ namespace Microsoft.CodeAnalysis.Editor.Implementation.IntelliSense
     internal interface IDocumentProvider
     {
         Document? GetDocument(ITextSnapshot snapshot, CancellationToken cancellationToken);
-    }
-
-    internal class DocumentProvider : IDocumentProvider
-    {
-        private readonly IThreadingContext _threadingContext;
-
-        public DocumentProvider(IThreadingContext threadingContext)
-        {
-            _threadingContext = threadingContext;
-        }
-
-        public Document? GetDocument(ITextSnapshot snapshot, CancellationToken cancellationToken)
-        {
-            _threadingContext.ThrowIfNotOnBackgroundThread();
-            return snapshot.AsText().GetDocumentWithFrozenPartialSemantics(cancellationToken);
-        }
     }
 }
